GCW-3160 Fix dashboard order links
